backend/requirements.txt
openai>=1.30.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.45.0
prophet>=1.1.5
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
mlflow>=2.13.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
openpyxl>=3.1.0
scipy>=1.13.0
pydantic>=2.7.0
httpx>=0.27.0
aiofiles>=23.2.0
